Made security updates to the servlets

+++ src/com/blueprintit/jspboard/servlets/FileUpload.java
@@ -7,6 +7,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.ServletException;
 import java.sql.Connection;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.io.IOException;
 import com.oreilly.servlet.MultipartRequest;
@@ -29,21 +30,31 @@ public class FileUpload extends HttpServlet
 		}
 		if ((redirect!=null)&&(message!=null))
 		{
-			Enumeration loop = req.getFileNames();
-			while (loop.hasMoreElements())
+			try
 			{
-				String fileid = loop.nextElement().toString();
-				String name = req.getOriginalFileName(fileid);
-				String filename = req.getFilesystemName(fileid);
-				String content = req.getContentType(fileid);
-				try
+				ResultSet results = conn.createStatement().executeQuery("SELECT Login.id FROM Login,Message WHERE Login.person=Message.owner AND Message.id="+message+" AND Login.id='"+request.getRemoteUser()+"';");
+				if ((request.isUserInRole("messageadmin"))||(results.next()))
 				{
-					conn.createStatement().executeUpdate("INSERT INTO File (name,filename,message,description,mimetype) VALUES ('"+name+"','"+filename+"',"+message+",'"+description+"','"+content+"');");
+					Enumeration loop = req.getFileNames();
+					while (loop.hasMoreElements())
+					{
+						String fileid = loop.nextElement().toString();
+						String name = req.getOriginalFileName(fileid);
+						String filename = req.getFilesystemName(fileid);
+						String content = req.getContentType(fileid);
+						try
+						{
+							conn.createStatement().executeUpdate("INSERT INTO File (name,filename,message,description,mimetype) VALUES ('"+name+"','"+filename+"',"+message+",'"+description+"','"+content+"');");
+						}
+						catch (SQLException e)
+						{
+							req.getFile(fileid).delete();
+						}
+					}
 				}
-				catch (SQLException e)
-				{
-					req.getFile(fileid).delete();
-				}
+			}
+			catch (Exception e)
+			{
 			}
 			request.getRequestDispatcher(redirect).forward(request,response);
 		}
